:bug: “wierd” values in “Prior ...” columns when there is no prior op
When there’s no prior operation the values for those columns are `nil`
this results in the cell value from the template being used. Because the
template contains values for all of the cells to make it easier to
visualize the result this results in “strang” values being copied over
for rows where the cells should otherwise have been left empty.

Copying the cell value from the template when the given value is `nil`
is a feature of the template library designed to aid in using formulas
(so that the formula can be propogated to new rows.)

To work around this I’ve changed the query so that `nil`s are replaced
with empty strings. (This won’t be an issue because there aren’t any
columns in the template where we _want_ the value copied.)